User feat/delete users (#162)
* feat: Add delete my account button to account settings page

* feat: add delete to user endpoints with constraints checking

* feat: add the ability to pass the id of entity owner to fetch entities of a specific owner

* fix: fix deleteItinerary condition in hasBookings function

* feat: add "delete" attribute to product schema and adjusted affected functions accordingly

* refactor: update comments for clarity and improve JSX formatting in ProductsModal component

* feat: add "delete" attribute and adjusted affected functions accordingly

* feat: update getAllActivities to accept filter parameters and generically filter based on the filter object

* feat: implement soft delete for itineraries and update fetching logic to support filtering

* feat: add bookingType and bookingStatus enums, and update related interfaces and constants

* feat: update getAllActivities to accept a filter parameter for improved activity retrieval

* fix: fix delete logic to avoid some bugs

* fix: added criteria that restricts the advertiser from deleting an activity with bookings

* fix: update deleteProduct to return the updated product after marking it as deleted

* refactor: remove unused IItinerary and IActivity interfaces from types definition

* refactor: remove unused bookingType and bookingStatus enums from types definition

* refactor: rename IQueryParamsFilter to IBookingQueryParamsFilter for clarity

* fix: Add activity date as the selected date for new created bookings

* refactor: refactored user-controller

* refactor: user Role enum in all possible places

* fix: adjusted filter logic for hasBookings function to check if an event has bookings or not

* refactor: removed console.log

* refactor: removed unused imports

* fix: updated filter logic in populateBookings to correctly handle ownerId and 3rd party bookings

* fix: modify getAllActivities to accept owner query parameters for filtering

* feat: update getAllProducts to accept query parameters for filtering by seller

* fix: remove "avail" query parameter since after inspection it was discovered to be redundant

* feat: Pop up a toast on account deletion errors

* fix: Add back the call to userService update by Id

* fix: updated filter Params to get correct seller products to be deleted

* fix: show seller name instead of seller Id

* fix: add constraints before deleting entities

* fix: update currency display from USD to EGP in product report

* fix: restore TripDetails route and add navigation component inside the TripDetails page

* fix: update cancelled booking status check in ActivityDetails and ItineraryDetails components

---------

+++ entertainment/src/services/activities-service.ts
@@ -1,10 +1,12 @@
 import type { IActivity } from "../database/models/Activity";
 import type { IRating } from "@/database/shared";
 import * as activitiesRepository from "../database/repositories/activities-repository";
+import { hasBookings } from "@/utils/booking-axios-instance";
+import { bookingType } from "@/utils/constants";
 
 // Get all activities
-export async function getAllActivities() {
-  return activitiesRepository.getAllActivities();
+export async function getAllActivities(filter: Partial<IActivity>) {
+  return activitiesRepository.getAllActivities(filter);
 }
 
 // Get all appropriate activities
@@ -33,9 +35,10 @@ export async function updateActivity(id: string, activitiesData: IActivity) {
 
 // Delete an activity
 export async function deleteActivity(id: string) {
-  return activitiesRepository.deleteActivity(id);
-}
-
-export async function getActivitiesByOwner(ownerId: string) {
-  return activitiesRepository.getActivitiesByOwner(ownerId);
+  const isBooked = await hasBookings(id, bookingType.Activity);
+  if (isBooked) {
+    throw new Error("Cannot delete activity with bookings");
+  } else {
+    return activitiesRepository.deleteActivity(id);
+  }
 }
